\name{RPqa-package}
\alias{RPqa-package}
\alias{RPqa}
\docType{package}
\title{
Recurrence Plot quantitative analysis
}
\description{
This package provides tools to create and quantitavely analysis recurrence plots, cross-recurrene plots and joint rcurrence plots.
}
\details{
\tabular{ll}{
Package: \tab RPqa\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-XX-XX\cr
License: \tab GPL-2\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Author: Nicola ZACCARELLI <nicola.zaccarelli@gmail.com>
Maintainer: Nicola ZACCARELLI <nicola.zaccarelli@gmail.com>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
